allow too many parent(/..) steps
This is mainly done by checking for some
undefs. Paths with too many parent steps
e.g. //../../CCC are useful to search
for keys (here: CCC) that are at least
n steps above a leaf.
The undefs occur because the // finds
lots of steps not all having that
many parents. We simply skip those and
therefore do not die and therefore
allow the use-case.