_any() saves a dereference
So far _any() returned an array by explicitely
dereferencing the array-ref after recursion.

We now directly return the array-ref and the
de-ref is done externally when needed.

This simplifies _any() for later refactoring
and even brings a slight speed-gain of ~1% in our
reference benchmarks. Probably more when _any()
would return lots of results in the benchmarks.