don't try to use warnings.pm inside safe container
Trying to load warnings.pm inside the Safe container does not actually
work, so the warnings code is not available inside the container. This
means warnings->unimport does not exist, but perl ignores the call.

Future versions of perl are intending to throw errors on calls like
this. Since we're trying to disable all warnings, we can just manually
set ${^WARNING_BITS} to an empty string.